+++ packages/lucid/test/specs/stake.ts
@@ -1,5 +1,6 @@
 import { Console, Effect, pipe, Schedule } from "effect";
 import { User } from "./services";
+import { handleSignSubmit, withLogRetry } from "./utils";
 
 export const registerStake = Effect.gen(function* ($) {
   const { user } = yield* User;
@@ -11,21 +12,15 @@ export const registerStake = Effect.gen(function* ($) {
     .newTx()
     .registerStake(rewardAddress)
     .completeProgram();
-  const signed = yield* signBuilder.sign.withWallet().completeProgram();
-  const txHash = yield* signed.submitProgram();
-  yield* Effect.promise(() => user.awaitTx(txHash, 20_000));
-  yield* Effect.logInfo(txHash);
+  return signBuilder;
 }).pipe(
-  Effect.tapErrorCause(Effect.log),
-  Effect.tapErrorCause(Console.log),
+  Effect.flatMap(handleSignSubmit),
   Effect.catchTag("TxSubmitError", (error) =>
     error.message.includes("StakeKeyAlreadyRegisteredDELEG")
       ? Effect.void
       : Effect.fail(error),
   ),
-  Effect.retry(
-    Schedule.compose(Schedule.exponential(20_000), Schedule.recurs(4)),
-  ),
+  withLogRetry,
 );
 
 export const deRegisterStake = Effect.gen(function* ($) {
@@ -39,17 +34,9 @@ export const deRegisterStake = Effect.gen(function* ($) {
     .newTx()
     .deRegisterStake(rewardAddress)
     .completeProgram();
-  const signed = yield* signBuilder.sign.withWallet().completeProgram();
-  const txHash = yield* signed.submitProgram();
-  yield* Effect.promise(() => user.awaitTx(txHash, 20_000));
-  yield* Effect.logInfo(txHash);
-}).pipe(
-  Effect.tapErrorCause(Effect.log),
-  Effect.tapErrorCause(Console.log),
-  Effect.retry(
-    Schedule.compose(Schedule.exponential(20_000), Schedule.recurs(4)),
-  ),
-);
+  return signBuilder;
+}).pipe(Effect.flatMap(handleSignSubmit), withLogRetry);
+
 export const registerDeregisterStake = Effect.gen(function* ($) {
   const { user } = yield* User;
   const rewardAddress = yield* pipe(
@@ -61,21 +48,15 @@ export const registerDeregisterStake = Effect.gen(function* ($) {
     .registerStake(rewardAddress)
     .deRegisterStake(rewardAddress)
     .completeProgram();
-  const signed = yield* signBuilder.sign.withWallet().completeProgram();
-  const txHash = yield* signed.submitProgram();
-  yield* Effect.promise(() => user.awaitTx(txHash, 20_000));
-  yield* Effect.logInfo(txHash);
+  return signBuilder;
 }).pipe(
-  Effect.tapErrorCause(Effect.log),
-  Effect.tapErrorCause(Console.log),
+  Effect.flatMap(handleSignSubmit),
   Effect.catchTag("TxSubmitError", (error) =>
     error.message.includes("StakeKeyAlreadyRegisteredDELEG")
       ? Effect.void
       : Effect.fail(error),
   ),
-  Effect.retry(
-    Schedule.compose(Schedule.exponential(20_000), Schedule.recurs(4)),
-  ),
+  withLogRetry,
 );
 
 export const withdrawZero = Effect.gen(function* ($) {
@@ -88,14 +69,5 @@ export const withdrawZero = Effect.gen(function* ($) {
     .newTx()
     .withdraw(rewardAddress, 0n)
     .completeProgram();
-  const signed = yield* signBuilder.sign.withWallet().completeProgram();
-  const txHash = yield* signed.submitProgram();
-  yield* Effect.promise(() => user.awaitTx(txHash, 20_000));
-  yield* Effect.logInfo(txHash);
-}).pipe(
-  Effect.tapErrorCause(Effect.log),
-  Effect.tapErrorCause(Console.log),
-  Effect.retry(
-    Schedule.compose(Schedule.exponential(20_000), Schedule.recurs(4)),
-  ),
-);
+  return signBuilder;
+}).pipe(Effect.flatMap(handleSignSubmit), withLogRetry);