+++ packages/provider/test/koios.test.ts
@@ -5,14 +5,15 @@ import {
   ProtocolParameters,
   UTxO,
 } from "@lucid-evolution/core-types";
-import { assert, describe, test } from "vitest";
+import { assert, describe, expect, test } from "vitest";
+import { expectedProtocolParameters } from "./protocolParameters";
 
 //TODO: improve test assetion
 describe.sequential("Koios", () => {
   const koios = new Koios("https://api.koios.rest/api/v1");
   test("getProtocolParameters", async () => {
     const pp: ProtocolParameters = await koios.getProtocolParameters();
-    assert(pp);
+    assert.deepEqual(pp, expectedProtocolParameters);
   });
 
   test("getUtxos", async () => {
@@ -70,7 +71,6 @@ describe.sequential("Koios", () => {
   });
 
   test("submitTxBadRequest", async () => {
-    const txId: string = await koios.submitTx("80");
-    assert(txId);
+    await expect(() => koios.submitTx("80")).rejects.toThrowError();
   });
 });